Renamed resources.py to apio_context.py. No behavior change.
