+++ apio/commands/boards.py
@@ -10,7 +10,7 @@
 from pathlib import Path
 from varname import nameof
 import click
-from apio.resources import ApioContext
+from apio.apio_context import ApioContext
 from apio import cmd_util
 from apio.commands import options
