+++ apio/commands/install.py
@@ -12,7 +12,7 @@ from typing import Tuple
 from varname import nameof
 import click
 from apio.managers.old_installer import Installer
-from apio.resources import ApioContext
+from apio.apio_context import ApioContext
 from apio import cmd_util
 from apio.commands import options
