+++ apio/commands/modify.py
@@ -13,7 +13,7 @@ import click
 from apio.managers.project import Project
 from apio import cmd_util
 from apio.commands import options
-from apio.resources import ApioContext
+from apio.apio_context import ApioContext
 
 
 # ---------------------------
