+++ apio/commands/test.py
@@ -12,7 +12,7 @@ import click
 from apio.managers.scons import SCons
 from apio import cmd_util
 from apio.commands import options
-from apio.resources import ApioContext
+from apio.apio_context import ApioContext
 
 
 # ---------------------------
